refactor(components): 컴포넌트 폴더 구조 변경
- nodes: 다른 컴포넌트를 사용하지 않는 가장 작은 단위의 컴포넌트
- trees: 다른 컴포넌트를 사용하는 그보다 큰 단위의 컴포넌트
-pages: 라우터에서 사용하는 페이지 단위의 컴포넌트
