+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export { default as Header } from "./Header";
-export { default as Product } from "./Product";
-export { default as ProductList } from "./ProductList";
+export { default as Header } from "./nodes/Header";
+export { default as Product } from "./nodes/Product";
+export { default as ProductList } from "./trees/ProductList";
